% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{AclEntry}
\alias{AclEntry}
\title{Cloud SQL Administration API Objects 
Creates and configures Cloud SQL instances, which provide fully-managed MySQL databases.}
\usage{
AclEntry(expirationTime = NULL, name = NULL, value = NULL)
}
\arguments{
\item{expirationTime}{The time when this access control entry expires in RFC 3339 format, for example 2012-11-15T16:19:00}

\item{name}{An optional label to identify this entry}

\item{value}{The whitelisted value for the access control list}
}
\value{
AclEntry object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:16:34
filename: /Users/mark/dev/R/autoGoogleAPI/googlesqladminv1beta4.auto/R/sqladmin_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AclEntry Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An entry for an Access Control list.
}
