% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{BackupConfiguration}
\alias{BackupConfiguration}
\title{BackupConfiguration Object}
\usage{
BackupConfiguration(binaryLogEnabled = NULL, enabled = NULL,
  startTime = NULL)
}
\arguments{
\item{binaryLogEnabled}{Whether binary log is enabled}

\item{enabled}{Whether this configuration is enabled}

\item{startTime}{Start time for the daily backup configuration in UTC timezone in the 24 hour format - HH:MM}
}
\value{
BackupConfiguration object
}
\description{
BackupConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance backup configuration.
}

