% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{BackupRun}
\alias{BackupRun}
\title{BackupRun Object}
\usage{
BackupRun(description = NULL, endTime = NULL, enqueuedTime = NULL,
  error = NULL, id = NULL, instance = NULL, selfLink = NULL,
  startTime = NULL, status = NULL, type = NULL, windowStartTime = NULL)
}
\arguments{
\item{description}{The description of this run, only applicable to on-demand backups}

\item{endTime}{The time the backup operation completed in UTC timezone in RFC 3339 format, for example 2012-11-15T16:19:00}

\item{enqueuedTime}{The time the run was enqueued in UTC timezone in RFC 3339 format, for example 2012-11-15T16:19:00}

\item{error}{Information about why the backup operation failed}

\item{id}{A unique identifier for this backup run}

\item{instance}{Name of the database instance}

\item{selfLink}{The URI of this resource}

\item{startTime}{The time the backup operation actually started in UTC timezone in RFC 3339 format, for example 2012-11-15T16:19:00}

\item{status}{The status of this run}

\item{type}{The type of this run; can be either 'AUTOMATED' or 'ON_DEMAND'}

\item{windowStartTime}{The start time of the backup window during which this the backup was attempted in RFC 3339 format, for example 2012-11-15T16:19:00}
}
\value{
BackupRun object
}
\description{
BackupRun Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A database instance backup run resource.
}
\seealso{
Other BackupRun functions: \code{\link{sql.backupRuns.insert}}
}
