% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{BinLogCoordinates}
\alias{BinLogCoordinates}
\title{BinLogCoordinates Object}
\usage{
BinLogCoordinates(binLogFileName = NULL, binLogPosition = NULL)
}
\arguments{
\item{binLogFileName}{Name of the binary log file for a Cloud SQL instance}

\item{binLogPosition}{Position (offset) within the binary log file}
}
\value{
BinLogCoordinates object
}
\description{
BinLogCoordinates Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Binary log coordinates.
}
