% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{DatabaseInstance.failoverReplica}
\alias{DatabaseInstance.failoverReplica}
\title{DatabaseInstance.failoverReplica Object}
\usage{
DatabaseInstance.failoverReplica(available = NULL, name = NULL)
}
\arguments{
\item{available}{The availability status of the failover replica}

\item{name}{The name of the failover replica}
}
\value{
DatabaseInstance.failoverReplica object
}
\description{
DatabaseInstance.failoverReplica Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The name and status of the failover replica. This property is applicable only to Second Generation instances.
}
\seealso{
Other DatabaseInstance functions: \code{\link{DatabaseInstance}},
  \code{\link{sql.instances.insert}},
  \code{\link{sql.instances.patch}},
  \code{\link{sql.instances.update}}
}

