% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{DatabaseInstance}
\alias{DatabaseInstance}
\title{DatabaseInstance Object}
\usage{
DatabaseInstance(DatabaseInstance.failoverReplica = NULL,
  backendType = NULL, currentDiskSize = NULL, databaseVersion = NULL,
  etag = NULL, failoverReplica = NULL, instanceType = NULL,
  ipAddresses = NULL, ipv6Address = NULL, masterInstanceName = NULL,
  maxDiskSize = NULL, name = NULL, onPremisesConfiguration = NULL,
  project = NULL, region = NULL, replicaConfiguration = NULL,
  replicaNames = NULL, selfLink = NULL, serverCaCert = NULL,
  serviceAccountEmailAddress = NULL, settings = NULL, state = NULL,
  suspensionReason = NULL)
}
\arguments{
\item{DatabaseInstance.failoverReplica}{The \link{DatabaseInstance.failoverReplica} object or list of objects}

\item{backendType}{FIRST_GEN: Basic Cloud SQL instance that runs in a Google-managed container}

\item{currentDiskSize}{The current disk usage of the instance in bytes}

\item{databaseVersion}{The database engine type and version}

\item{etag}{HTTP 1}

\item{failoverReplica}{The name and status of the failover replica}

\item{instanceType}{The instance type}

\item{ipAddresses}{The assigned IP addresses for the instance}

\item{ipv6Address}{The IPv6 address assigned to the instance}

\item{masterInstanceName}{The name of the instance which will act as master in the replication setup}

\item{maxDiskSize}{The maximum disk size of the instance in bytes}

\item{name}{Name of the Cloud SQL instance}

\item{onPremisesConfiguration}{Configuration specific to on-premises instances}

\item{project}{The project ID of the project containing the Cloud SQL instance}

\item{region}{The geographical region}

\item{replicaConfiguration}{Configuration specific to read-replicas replicating from on-premises masters}

\item{replicaNames}{The replicas of the instance}

\item{selfLink}{The URI of this resource}

\item{serverCaCert}{SSL configuration}

\item{serviceAccountEmailAddress}{The service account email address assigned to the instance}

\item{settings}{The user settings}

\item{state}{The current serving state of the Cloud SQL instance}

\item{suspensionReason}{If the instance state is SUSPENDED, the reason for the suspension}
}
\value{
DatabaseInstance object
}
\description{
DatabaseInstance Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Cloud SQL instance resource.
}
\seealso{
Other DatabaseInstance functions: \code{\link{DatabaseInstance.failoverReplica}},
  \code{\link{sql.instances.insert}},
  \code{\link{sql.instances.patch}},
  \code{\link{sql.instances.update}}
}

