% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{DatabasesListResponse}
\alias{DatabasesListResponse}
\title{DatabasesListResponse Object}
\usage{
DatabasesListResponse(items = NULL)
}
\arguments{
\item{items}{List of database resources in the instance}
}
\value{
DatabasesListResponse object
}
\description{
DatabasesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database list response.
}

