% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{ExportContext.csvExportOptions}
\alias{ExportContext.csvExportOptions}
\title{ExportContext.csvExportOptions Object}
\usage{
ExportContext.csvExportOptions(selectQuery = NULL)
}
\arguments{
\item{selectQuery}{The select query used to extract the data}
}
\value{
ExportContext.csvExportOptions object
}
\description{
ExportContext.csvExportOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Options for exporting data as CSV.
}
\seealso{
Other ExportContext functions: \code{\link{ExportContext.sqlExportOptions}},
  \code{\link{ExportContext}}
}

