% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{ExportContext}
\alias{ExportContext}
\title{ExportContext Object}
\usage{
ExportContext(ExportContext.csvExportOptions = NULL,
  ExportContext.sqlExportOptions = NULL, csvExportOptions = NULL,
  databases = NULL, fileType = NULL, sqlExportOptions = NULL,
  uri = NULL)
}
\arguments{
\item{ExportContext.csvExportOptions}{The \link{ExportContext.csvExportOptions} object or list of objects}

\item{ExportContext.sqlExportOptions}{The \link{ExportContext.sqlExportOptions} object or list of objects}

\item{csvExportOptions}{Options for exporting data as CSV}

\item{databases}{Databases (for example, guestbook) from which the export is made}

\item{fileType}{The file type for the specified uri}

\item{sqlExportOptions}{Options for exporting data as SQL statements}

\item{uri}{The path to the file in Google Cloud Storage where the export will be stored}
}
\value{
ExportContext object
}
\description{
ExportContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance export context.
}
\seealso{
Other ExportContext functions: \code{\link{ExportContext.csvExportOptions}},
  \code{\link{ExportContext.sqlExportOptions}}
}
