% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{ExportContext.sqlExportOptions}
\alias{ExportContext.sqlExportOptions}
\title{ExportContext.sqlExportOptions Object}
\usage{
ExportContext.sqlExportOptions(schemaOnly = NULL, tables = NULL)
}
\arguments{
\item{schemaOnly}{Export only schemas}

\item{tables}{Tables to export, or that were exported, from the specified database}
}
\value{
ExportContext.sqlExportOptions object
}
\description{
ExportContext.sqlExportOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Options for exporting data as SQL statements.
}
\seealso{
Other ExportContext functions: \code{\link{ExportContext.csvExportOptions}},
  \code{\link{ExportContext}}
}
