% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{FailoverContext}
\alias{FailoverContext}
\title{FailoverContext Object}
\usage{
FailoverContext(settingsVersion = NULL)
}
\arguments{
\item{settingsVersion}{The current settings version of this instance}
}
\value{
FailoverContext object
}
\description{
FailoverContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance failover context.
}

