% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{Flag}
\alias{Flag}
\title{Flag Object}
\usage{
Flag(allowedStringValues = NULL, appliesTo = NULL, maxValue = NULL,
  minValue = NULL, name = NULL, requiresRestart = NULL, type = NULL)
}
\arguments{
\item{allowedStringValues}{For STRING flags, a list of strings that the value can be set to}

\item{appliesTo}{The database version this flag applies to}

\item{maxValue}{For INTEGER flags, the maximum allowed value}

\item{minValue}{For INTEGER flags, the minimum allowed value}

\item{name}{This is the name of the flag}

\item{requiresRestart}{Indicates whether changing this flag will trigger a database restart}

\item{type}{The type of the flag}
}
\value{
Flag object
}
\description{
Flag Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Google Cloud SQL service flag resource.
}
