% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{FlagsListResponse}
\alias{FlagsListResponse}
\title{FlagsListResponse Object}
\usage{
FlagsListResponse(items = NULL)
}
\arguments{
\item{items}{List of flags}
}
\value{
FlagsListResponse object
}
\description{
FlagsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Flags list response.
}
