% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{ImportContext.csvImportOptions}
\alias{ImportContext.csvImportOptions}
\title{ImportContext.csvImportOptions Object}
\usage{
ImportContext.csvImportOptions(columns = NULL, table = NULL)
}
\arguments{
\item{columns}{The columns to which CSV data is imported}

\item{table}{The table to which CSV data is imported}
}
\value{
ImportContext.csvImportOptions object
}
\description{
ImportContext.csvImportOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Options for importing data as CSV.
}
\seealso{
Other ImportContext functions: \code{\link{ImportContext}}
}
