% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{ImportContext}
\alias{ImportContext}
\title{ImportContext Object}
\usage{
ImportContext(ImportContext.csvImportOptions = NULL,
  csvImportOptions = NULL, database = NULL, fileType = NULL, uri = NULL)
}
\arguments{
\item{ImportContext.csvImportOptions}{The \link{ImportContext.csvImportOptions} object or list of objects}

\item{csvImportOptions}{Options for importing data as CSV}

\item{database}{The database (for example, guestbook) to which the import is made}

\item{fileType}{The file type for the specified uri}

\item{uri}{A path to the file in Google Cloud Storage from which the import is made}
}
\value{
ImportContext object
}
\description{
ImportContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance import context.
}
\seealso{
Other ImportContext functions: \code{\link{ImportContext.csvImportOptions}}
}
