% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{InstancesExportRequest}
\alias{InstancesExportRequest}
\title{InstancesExportRequest Object}
\usage{
InstancesExportRequest(exportContext = NULL)
}
\arguments{
\item{exportContext}{Contains details about the export operation}
}
\value{
InstancesExportRequest object
}
\description{
InstancesExportRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance export request.
}
\seealso{
Other InstancesExportRequest functions: \code{\link{sql.instances.export}}
}
