% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{InstancesFailoverRequest}
\alias{InstancesFailoverRequest}
\title{InstancesFailoverRequest Object}
\usage{
InstancesFailoverRequest(failoverContext = NULL)
}
\arguments{
\item{failoverContext}{Failover Context}
}
\value{
InstancesFailoverRequest object
}
\description{
InstancesFailoverRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Instance failover request.
}
\seealso{
Other InstancesFailoverRequest functions: \code{\link{sql.instances.failover}}
}

