% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{InstancesListResponse}
\alias{InstancesListResponse}
\title{InstancesListResponse Object}
\usage{
InstancesListResponse(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{List of database instance resources}

\item{nextPageToken}{The continuation token, used to page through large result sets}
}
\value{
InstancesListResponse object
}
\description{
InstancesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instances list response.
}

