% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{InstancesTruncateLogRequest}
\alias{InstancesTruncateLogRequest}
\title{InstancesTruncateLogRequest Object}
\usage{
InstancesTruncateLogRequest(truncateLogContext = NULL)
}
\arguments{
\item{truncateLogContext}{Contains details about the truncate log operation}
}
\value{
InstancesTruncateLogRequest object
}
\description{
InstancesTruncateLogRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Instance truncate log request.
}
\seealso{
Other InstancesTruncateLogRequest functions: \code{\link{sql.instances.truncateLog}}
}
