% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{Labels}
\alias{Labels}
\title{Labels Object}
\usage{
Labels(key = NULL, value = NULL)
}
\arguments{
\item{key}{The key of the label}

\item{value}{The value of the label}
}
\value{
Labels object
}
\description{
Labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
User defined labels for Cloud SQL instances.
}
