% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{MySqlReplicaConfiguration}
\alias{MySqlReplicaConfiguration}
\title{MySqlReplicaConfiguration Object}
\usage{
MySqlReplicaConfiguration(caCertificate = NULL, clientCertificate = NULL,
  clientKey = NULL, connectRetryInterval = NULL, dumpFilePath = NULL,
  masterHeartbeatPeriod = NULL, password = NULL, sslCipher = NULL,
  username = NULL, verifyServerCertificate = NULL)
}
\arguments{
\item{caCertificate}{PEM representation of the trusted CA's x509 certificate}

\item{clientCertificate}{PEM representation of the slave's x509 certificate}

\item{clientKey}{PEM representation of the slave's private key}

\item{connectRetryInterval}{Seconds to wait between connect retries}

\item{dumpFilePath}{Path to a SQL dump file in Google Cloud Storage from which the slave instance is to be created}

\item{masterHeartbeatPeriod}{Interval in milliseconds between replication heartbeats}

\item{password}{The password for the replication connection}

\item{sslCipher}{A list of permissible ciphers to use for SSL encryption}

\item{username}{The username for the replication connection}

\item{verifyServerCertificate}{Whether or not to check the master's Common Name value in the certificate that it sends during the SSL handshake}
}
\value{
MySqlReplicaConfiguration object
}
\description{
MySqlReplicaConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Read-replica configuration specific to MySQL databases.
}

