% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{OnPremisesConfiguration}
\alias{OnPremisesConfiguration}
\title{OnPremisesConfiguration Object}
\usage{
OnPremisesConfiguration(hostPort = NULL)
}
\arguments{
\item{hostPort}{The host and port of the on-premises instance in host:port format}
}
\value{
OnPremisesConfiguration object
}
\description{
OnPremisesConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
On-premises instance configuration.
}
