% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{Operation}
\alias{Operation}
\title{Operation Object}
\usage{
Operation(endTime = NULL, error = NULL, exportContext = NULL,
  importContext = NULL, insertTime = NULL, name = NULL,
  operationType = NULL, selfLink = NULL, startTime = NULL,
  status = NULL, targetId = NULL, targetLink = NULL,
  targetProject = NULL, user = NULL)
}
\arguments{
\item{endTime}{The time this operation finished in UTC timezone in RFC 3339 format, for example 2012-11-15T16:19:00}

\item{error}{If errors occurred during processing of this operation, this field will be populated}

\item{exportContext}{The context for export operation, if applicable}

\item{importContext}{The context for import operation, if applicable}

\item{insertTime}{The time this operation was enqueued in UTC timezone in RFC 3339 format, for example 2012-11-15T16:19:00}

\item{name}{An identifier that uniquely identifies the operation}

\item{operationType}{The type of the operation}

\item{selfLink}{The URI of this resource}

\item{startTime}{The time this operation actually started in UTC timezone in RFC 3339 format, for example 2012-11-15T16:19:00}

\item{status}{The status of an operation}

\item{targetId}{Name of the database instance related to this operation}

\item{targetLink}{The URI of the instance related to the operation}

\item{targetProject}{The project ID of the target instance related to this operation}

\item{user}{The email address of the user who initiated this operation}
}
\value{
Operation object
}
\description{
Operation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An Operations resource contains information about database instance operations such as create, delete, and restart. Operations resources are created in response to operations that were initiated; you never create them directly.
}

