% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{OperationError}
\alias{OperationError}
\title{OperationError Object}
\usage{
OperationError(code = NULL, message = NULL)
}
\arguments{
\item{code}{Identifies the specific error that occurred}

\item{message}{Additional information about the error encountered}
}
\value{
OperationError object
}
\description{
OperationError Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance operation error.
}

