% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{OperationErrors}
\alias{OperationErrors}
\title{OperationErrors Object}
\usage{
OperationErrors(errors = NULL)
}
\arguments{
\item{errors}{The list of errors encountered while processing this operation}
}
\value{
OperationErrors object
}
\description{
OperationErrors Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance operation errors list wrapper.
}
