% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{OperationsListResponse}
\alias{OperationsListResponse}
\title{OperationsListResponse Object}
\usage{
OperationsListResponse(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{List of operation resources}

\item{nextPageToken}{The continuation token, used to page through large result sets}
}
\value{
OperationsListResponse object
}
\description{
OperationsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance list operations response.
}

