% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{ReplicaConfiguration}
\alias{ReplicaConfiguration}
\title{ReplicaConfiguration Object}
\usage{
ReplicaConfiguration(failoverTarget = NULL,
  mysqlReplicaConfiguration = NULL)
}
\arguments{
\item{failoverTarget}{Specifies if the replica is the failover target}

\item{mysqlReplicaConfiguration}{MySQL specific configuration when replicating from a MySQL on-premises master}
}
\value{
ReplicaConfiguration object
}
\description{
ReplicaConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Read-replica configuration for connecting to the master.
}
