% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{RestoreBackupContext}
\alias{RestoreBackupContext}
\title{RestoreBackupContext Object}
\usage{
RestoreBackupContext(backupRunId = NULL, instanceId = NULL)
}
\arguments{
\item{backupRunId}{The ID of the backup run to restore from}

\item{instanceId}{The ID of the instance that the backup was taken from}
}
\value{
RestoreBackupContext object
}
\description{
RestoreBackupContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance restore from backup context.
}
