% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{Settings}
\alias{Settings}
\title{Settings Object}
\usage{
Settings(activationPolicy = NULL, authorizedGaeApplications = NULL,
  availabilityType = NULL, backupConfiguration = NULL,
  crashSafeReplicationEnabled = NULL, dataDiskSizeGb = NULL,
  dataDiskType = NULL, databaseFlags = NULL,
  databaseReplicationEnabled = NULL, ipConfiguration = NULL,
  labels = NULL, locationPreference = NULL, maintenanceWindow = NULL,
  pricingPlan = NULL, replicationType = NULL, settingsVersion = NULL,
  storageAutoResize = NULL, storageAutoResizeLimit = NULL, tier = NULL)
}
\arguments{
\item{activationPolicy}{The activation policy specifies when the instance is activated; it is applicable only when the instance state is RUNNABLE}

\item{authorizedGaeApplications}{The App Engine app IDs that can access this instance}

\item{availabilityType}{Reserved for future use}

\item{backupConfiguration}{The daily backup configuration for the instance}

\item{crashSafeReplicationEnabled}{Configuration specific to read replica instances}

\item{dataDiskSizeGb}{The size of data disk, in GB}

\item{dataDiskType}{The type of data disk}

\item{databaseFlags}{The database flags passed to the instance at startup}

\item{databaseReplicationEnabled}{Configuration specific to read replica instances}

\item{ipConfiguration}{The settings for IP Management}

\item{labels}{User defined labels}

\item{locationPreference}{The location preference settings}

\item{maintenanceWindow}{The maintenance window for this instance}

\item{pricingPlan}{The pricing plan for this instance}

\item{replicationType}{The type of replication this instance uses}

\item{settingsVersion}{The version of instance settings}

\item{storageAutoResize}{Configuration to increase storage size automatically}

\item{storageAutoResizeLimit}{The maximum size to which storage capacity can be automatically increased}

\item{tier}{The tier of service for this instance, for example D1, D2}
}
\value{
Settings object
}
\description{
Settings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database instance settings.
}
