% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_functions.R
\name{sql.backupRuns.insert}
\alias{sql.backupRuns.insert}
\title{Creates a new backup run on demand. This method is applicable only to Second Generation instances.}
\usage{
sql.backupRuns.insert(BackupRun, project, instance)
}
\arguments{
\item{BackupRun}{The \link{BackupRun} object to pass to this method}

\item{project}{Project ID of the project that contains the instance}

\item{instance}{Cloud SQL instance ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/sql/docs/reference/latest}{Google Documentation}

Other BackupRun functions: \code{\link{BackupRun}}
}

