% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_functions.R
\name{sql.backupRuns.list}
\alias{sql.backupRuns.list}
\title{Lists all backup runs associated with a given instance and configuration in the reverse chronological order of the enqueued time.}
\usage{
sql.backupRuns.list(project, instance, maxResults = NULL, pageToken = NULL)
}
\arguments{
\item{project}{Project ID of the project that contains the instance}

\item{instance}{Cloud SQL instance ID}

\item{maxResults}{Maximum number of backup runs per response}

\item{pageToken}{A previously-returned page token representing part of the larger set of results to view}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/sqlservice.admin
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/sqlservice.admin)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/sql/docs/reference/latest}{Google Documentation}
}
