% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_functions.R
\name{sql.instances.export}
\alias{sql.instances.export}
\title{Exports data from a Cloud SQL instance to a Google Cloud Storage bucket as a MySQL dump file.}
\usage{
sql.instances.export(InstancesExportRequest, project, instance)
}
\arguments{
\item{InstancesExportRequest}{The \link{InstancesExportRequest} object to pass to this method}

\item{project}{Project ID of the project that contains the instance to be exported}

\item{instance}{Cloud SQL instance ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/sql/docs/reference/latest}{Google Documentation}

Other InstancesExportRequest functions: \code{\link{InstancesExportRequest}}
}

