% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_functions.R
\name{sql.instances.import}
\alias{sql.instances.import}
\title{Imports data into a Cloud SQL instance from a MySQL dump file in Google Cloud Storage.}
\usage{
sql.instances.import(InstancesImportRequest, project, instance)
}
\arguments{
\item{InstancesImportRequest}{The \link{InstancesImportRequest} object to pass to this method}

\item{project}{Project ID of the project that contains the instance}

\item{instance}{Cloud SQL instance ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/sql/docs/reference/latest}{Google Documentation}

Other InstancesImportRequest functions: \code{\link{InstancesImportRequest}}
}
