% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_functions.R
\name{sql.sslCerts.createEphemeral}
\alias{sql.sslCerts.createEphemeral}
\title{Generates a short-lived X509 certificate containing the provided public key and signed by a private key specific to the target instance. Users may use the certificate to authenticate as themselves when connecting to the database.}
\usage{
sql.sslCerts.createEphemeral(SslCertsCreateEphemeralRequest, project, instance)
}
\arguments{
\item{SslCertsCreateEphemeralRequest}{The \link{SslCertsCreateEphemeralRequest} object to pass to this method}

\item{project}{Project ID of the Cloud SQL project}

\item{instance}{Cloud SQL instance ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/sqlservice.admin
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/sqlservice.admin)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/sql/docs/reference/latest}{Google Documentation}

Other SslCertsCreateEphemeralRequest functions: \code{\link{SslCertsCreateEphemeralRequest}}
}

