% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_functions.R
\name{sql.sslCerts.delete}
\alias{sql.sslCerts.delete}
\title{Deletes the SSL certificate. The change will not take effect until the instance is restarted.}
\usage{
sql.sslCerts.delete(project, instance, sha1Fingerprint)
}
\arguments{
\item{project}{Project ID of the project that contains the instance to be deleted}

\item{instance}{Cloud SQL instance ID}

\item{sha1Fingerprint}{Sha1 FingerPrint}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/sqlservice.admin
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/sqlservice.admin)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/sql/docs/reference/latest}{Google Documentation}
}
