% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{SslCert}
\alias{SslCert}
\title{SslCert Object}
\usage{
SslCert(cert = NULL, certSerialNumber = NULL, commonName = NULL,
  createTime = NULL, expirationTime = NULL, instance = NULL,
  selfLink = NULL, sha1Fingerprint = NULL)
}
\arguments{
\item{cert}{PEM representation}

\item{certSerialNumber}{Serial number, as extracted from the certificate}

\item{commonName}{User supplied name}

\item{createTime}{The time when the certificate was created in RFC 3339 format, for example 2012-11-15T16:19:00}

\item{expirationTime}{The time when the certificate expires in RFC 3339 format, for example 2012-11-15T16:19:00}

\item{instance}{Name of the database instance}

\item{selfLink}{The URI of this resource}

\item{sha1Fingerprint}{Sha1 Fingerprint}
}
\value{
SslCert object
}
\description{
SslCert Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
SslCerts Resource
}
