% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{SslCertsCreateEphemeralRequest}
\alias{SslCertsCreateEphemeralRequest}
\title{SslCertsCreateEphemeralRequest Object}
\usage{
SslCertsCreateEphemeralRequest(public_key = NULL)
}
\arguments{
\item{public_key}{PEM encoded public key to include in the signed certificate}
}
\value{
SslCertsCreateEphemeralRequest object
}
\description{
SslCertsCreateEphemeralRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
SslCerts create ephemeral certificate request.
}
\seealso{
Other SslCertsCreateEphemeralRequest functions: \code{\link{sql.sslCerts.createEphemeral}}
}

