% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{SslCertsInsertResponse}
\alias{SslCertsInsertResponse}
\title{SslCertsInsertResponse Object}
\usage{
SslCertsInsertResponse(clientCert = NULL, operation = NULL,
  serverCaCert = NULL)
}
\arguments{
\item{clientCert}{The new client certificate and private key}

\item{operation}{The operation to track the ssl certs insert request}

\item{serverCaCert}{The server Certificate Authority's certificate}
}
\value{
SslCertsInsertResponse object
}
\description{
SslCertsInsertResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
SslCert insert response.
}

