% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{TruncateLogContext}
\alias{TruncateLogContext}
\title{TruncateLogContext Object}
\usage{
TruncateLogContext(logType = NULL)
}
\arguments{
\item{logType}{The type of log to truncate}
}
\value{
TruncateLogContext object
}
\description{
TruncateLogContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Database Instance truncate log context.
}
