% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{User}
\alias{User}
\title{User Object}
\usage{
User(etag = NULL, host = NULL, instance = NULL, name = NULL,
  password = NULL, project = NULL)
}
\arguments{
\item{etag}{HTTP 1}

\item{host}{The host name from which the user can connect}

\item{instance}{The name of the Cloud SQL instance}

\item{name}{The name of the user in the Cloud SQL instance}

\item{password}{The password for the user}

\item{project}{The project ID of the project containing the Cloud SQL database}
}
\value{
User object
}
\description{
User Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Cloud SQL user resource.
}
\seealso{
Other User functions: \code{\link{sql.users.insert}},
  \code{\link{sql.users.update}}
}
