% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqladmin_objects.R
\name{UsersListResponse}
\alias{UsersListResponse}
\title{UsersListResponse Object}
\usage{
UsersListResponse(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{List of user resources in the instance}

\item{nextPageToken}{An identifier that uniquely identifies the operation}
}
\value{
UsersListResponse object
}
\description{
UsersListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
User list response.
}
